requirements.txt
Django==2.2
djangorestframework==3.9.0
drf-yasg==1.15.0